database/requirements.txt
neo4j==5.19.0
python-dotenv==1.0.1
requests==2.31.0